split out sql module
